Fixed tests and variable naming bugs.
